+++ vSMR/src/radar/RadarScreen.FlightDataAndCommands.cpp
@@ -250,7 +250,7 @@ void CSMRRadar::OnFlightPlanDisconnect(CFlightPlan FlightPlan)
 	TagLeaderLineLength.erase(callsign);
 	tagAreas.erase(callsign);
 	tagCollisionAreas.erase(callsign);
-	previousTagSize.erase(callsign);
+	DetailedTagCallsigns.erase(callsign);
 	TagDragOffsetFromCenter.erase(callsign);
 	RecentlyAutoMovedTags.erase(callsign);
 }